Fix training pipeline: corrected PM25Model initialization and dtype issues
- Fixed PM25Model constructor call in train.py (config object vs individual params)
- Fixed tensor dtype issues (float32 vs float64) for lead_times
- Updated config manager for multi-horizon temporal configuration
- Enhanced dataloader with proper variable handling
- Added debugging support in run.sh
- Fixed SLURM submit script for different GPU configurations
- Training successfully runs on LUMI with 4-8 GPUs

+++ src/model.py
@@ -77,7 +77,7 @@ class PM25Model(nn.Module):
     def forward(self, x, lead_times):
         B,C,H,W = x.shape
         device = x.device
-        lt = torch.tensor(lead_times, device=device)
+        lt = torch.tensor(lead_times, dtype=torch.float32, device=device)
         feats = checkpoint(self.climax.module.forward_encoder, x, lt, self.variables) if hasattr(self.climax, "module") else checkpoint(self.climax.forward_encoder, x, lt, self.variables)
         for blk in (self.climax.module.blocks if hasattr(self.climax, "module") else self.climax.blocks):
             feats = blk(feats)
